premier validation

+++ backend/Controllers/books.js
@@ -10,7 +10,7 @@ exports.viewBooks = (req, res, next) => {
 
 exports.viewBook = (req, res, next) => {
     Book.findOne({ _id: req.params.id })
-        .then(books => res.status(200).json(books))
+        .then(book => res.status(200).json(book))
         .catch(error => res.status(400).json({ error }));
 }
 
@@ -26,7 +26,7 @@ exports.addBook = (req, res, next) => {
    });
 
    book.save()
-   .then(() => res.status(201).json({ message: 'Objet enregistré !' }))
+   .then(() => res.status(201).json({ message: 'Livre enregistré !' }))
    .catch(error => res.status(400).json({ error }));
 
 }
@@ -45,7 +45,7 @@ exports.updateBook = (req, res, next) => {
     Book.findOne({_id: req.params.id})
         .then((book) => {
             if (book.userId != req.auth.user_Id) {
-                res.status(401).json({ message : 'Not authorized'});
+                res.status(403).json({ message : 'unauthorized request'});
             } else {
                 const filename = book.imageUrl.split('/images/')[1];
                 console.log(filename);
@@ -54,7 +54,7 @@ exports.updateBook = (req, res, next) => {
                 // Pour limiter la taille du server on supprime l'image qui n'est plus utilisé
                 fs.unlink(`images/${filename}`, () => {
                     Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
-                    .then(() => res.status(200).json({message : 'Objet modifié!'}))
+                    .then(() => res.status(200).json({message : 'Livre modifié!'}))
                     .catch(error => res.status(401).json({ error }));    });
               
             }
@@ -68,16 +68,16 @@ exports.updateBook = (req, res, next) => {
 exports.deleteBook = (req, res, next) => {
     Book.findOne({ _id: req.params.id})
         .then(book => {
-            // if (book.userId != req.auth.user_Id) {
-            //     res.status(401).json({message: 'Not authorized'});
-            // } else {
+            if (book.userId != req.auth.user_Id) {
+                res.status(403).json({ message : 'unauthorized request'});
+            } else {
                 const filename = book.imageUrl.split('/images/')[1];
                 fs.unlink(`./images/${filename}`, () => {
                     Book.deleteOne({_id: req.params.id})
-                        .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                        .then(() =>  res.status(200).json({message: 'Livre supprimé !'}))
                         .catch(error => res.status(401).json({ error }));
                 });
-            // }
+         }
         })
         .catch( error => {
             res.status(500).json({ error });
@@ -89,7 +89,7 @@ exports.deleteBook = (req, res, next) => {
         Book.findOne({ _id: req.params.id })
           .then((book) => {
             if (book.ratings.userId === req.auth.user_Id) {
-              res.status(401).json({ message: 'Not authorized' });
+                res.status(403).json({ message : 'unauthorized request'});
             } else {
              
               // On ajoute la nouvelle note et la nouvelle note moyenne
